Fix to enable only the owner of the image to edit/delete the image There is no check on the details of the user when the image is edited/deleted. This change allows only owner to edit/delete and display error message if the non-owner of the image try to edit/delete the image.

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,13 +92,21 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession httpSession) {
         Image image = imageService.getImage(imageId);
-
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
+
+        User loggeduser = (User) httpSession.getAttribute("loggeduser");
+
+        if(image.getUser().getUsername().equals(loggeduser.getUsername())){
+            String tags = convertTagsToString(image.getTags());
+            model.addAttribute("tags", tags);
+            return "images/edit";
+        } else{
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("editError","Only the owner of the image can edit the image");
+            return "images/image";
+        }
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -140,9 +148,19 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession httpSession) {
+        Image image = imageService.getImage(imageId);
+        User loggeduser = (User) httpSession.getAttribute("loggeduser");
+
+        if(image.getUser().getUsername().equals(loggeduser.getUsername())){
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        } else{
+            model.addAttribute("image", image);
+            model.addAttribute("tags", image.getTags());
+            model.addAttribute("deleteError","Only the owner of the image can delete the image");
+            return "images/image";
+        }
     }
